Folder and file structure and naming changes
