+++ src/index.js
@@ -3,7 +3,7 @@ import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { ConnectedRouter } from 'react-router-redux';
 import store, { history } from './store';
-import App from './containers/app';
+import App from './containers/App';
 
 import 'sanitize.css/sanitize.css';
 import './index.css';
